refactor(sync): channel-based structure, split oversized modules
