+++ src/daemon/server.ts
@@ -1,9 +1,9 @@
 import { createServer, type IncomingMessage, type Server, type ServerResponse } from 'node:http';
 import { type McpServer } from '@modelcontextprotocol/sdk/server/mcp.js';
 import { type MemoryClient } from '../lib/memory-client.js';
-import { type SyncResult } from '../sync/cycle.js';
+import { type SyncResult } from '../sync/git/cycle.js';
 import { type CouchSyncResult } from '../sync/couch/manager.js';
-import { type SyncStatus } from '../sync/manager.js';
+import { type SyncStatus } from '../sync/git/manager.js';
 import { dispatchMemory, isMemoryVerb, type MemoryVerb } from './actions.js';
 import { isAllowedHost, isAllowedOrigin, loopbackHosts } from './guards.js';
 import { handleMcp } from './mcp-http.js';
