+++ src/sync/git/manager.test.ts
