+++ src/sync/git/manager.ts
@@ -1,7 +1,7 @@
 import { type VaultsConfig } from '@agentage/memory-core';
-import { loadVaultsConfig } from '../lib/vaults.js';
-import { type CouchTargetStatus } from './couch/manager.js';
-import { type DiscoverStatus } from './discover/watcher.js';
+import { loadVaultsConfig } from '../../lib/vaults.js';
+import { type CouchTargetStatus } from '../couch/manager.js';
+import { type DiscoverStatus } from '../discover/watcher.js';
 import { runSyncCycle, type SyncResult } from './cycle.js';
 import { autoSyncTargets, intervalMs, syncTargets, type SyncTarget } from './planner.js';
